feat(pages): Scaffold second step, make StepperContentLayout only depend on itself
