+++ apps/chocolate-frontend/src/pages/Verify/StepperContentLayout.tsx
@@ -1,30 +1,51 @@
 import { Button } from '@mantine/core';
+import styled from 'styled-components';
 
-export enum FontWeights {
-  bold = 500,
-}
 interface VerifyFirstStepProps {
   hasNext?: boolean;
   hasPrev?: boolean;
   nextStep: VoidFunction;
   prevStep: VoidFunction;
+  className?: string;
 }
-export function StepperContentLayout(
+function StepperContentLayout(
   props: React.PropsWithChildren<VerifyFirstStepProps>
 ) {
-  const { nextStep, prevStep, hasNext, hasPrev, children } = props;
+  const { nextStep, prevStep, hasNext, hasPrev, children, ...rest } = props;
 
   return (
-    <>
+    <div {...rest}>
       <div className="StepsBody">{children}</div>
       <div className="StepsControl">
         {hasPrev && (
-          <Button variant="default" onClick={prevStep}>
+          <Button variant="default" onClick={prevStep} className="PrevStep">
             Back
           </Button>
         )}
-        {hasNext && <Button onClick={nextStep}>Next step</Button>}
+        {hasNext && (
+          <Button onClick={nextStep} className="NextStep">
+            Next step
+          </Button>
+        )}
       </div>
-    </>
+    </div>
   );
 }
+
+export default styled(StepperContentLayout)`
+  width: 100%;
+  height: 100%;
+  display: flex;
+  flex-direction: column;
+
+  .StepsControl {
+    display: grid;
+    grid-template-columns: repeat(1, min-content);
+    grid-column-gap: 10px;
+    justify-content: space-between;
+
+    .NextStep {
+      grid-column: 2;
+    }
+  }
+`;
