+++ apps/chocolate-frontend/src/pages/Verify/types.ts
@@ -1,0 +1,6 @@
+export interface VerifyLayoutProps {
+  className?: string;
+}
+export interface ActiveMap {
+  [pageId: number]: boolean;
+}
